app: Fixed around orders scroll; Compute total sum on progressbar

+++ InitiativaRomania/app/src/main/java/com/initiativaromania/hartabanilorpublici/IRData/Buyer.java
@@ -12,5 +12,5 @@ public class Buyer implements Serializable {
     public double longitude;
     public double latitude;
 
-    public long totalPrice;
+    public double totalPrice;
 }
